+++ dashboard/api/src/main/java/com/dashboard/api/Repository/WidgetRepository.java
@@ -3,6 +3,6 @@ package com.dashboard.api.Repository;
 import org.springframework.data.repository.CrudRepository;
 import com.dashboard.api.Entity.Widget;
 
-public interface WidgetRepository extends CrudRepository<Widget, Integer> {
+public interface WidgetRepository extends CrudRepository<Widget, Long> {
 
 }
